+++ src/main/java/woo/ba/ben/bean/FloatPropertyAccessor.java
@@ -2,9 +2,9 @@ package woo.ba.ben.bean;
 
 
 public interface FloatPropertyAccessor extends IPropertyAccessor {
-    double get(Object bean, String field);
-    double getArrayElementAt(Object bean, String field, int index);
+    float get(Object bean, String field);
+    float getArrayElementAt(Object bean, String field, int index);
 
-    void set(Object bean, String field, double value);
-    void setArrayElementAt(Object bean, String field, int index, double value);
+    void set(Object bean, String field, float value);
+    void setArrayElementAt(Object bean, String field, int index, float value);
 }
